chore: move endpoint to github.installApp instead of workspace.installGithub
