+++ svc/api/openapi/openapi-split.yaml
@@ -124,8 +124,8 @@ tags:
     description: Customer Portal session management
   - name: ratelimit
     description: Rate limiting operations
-  - name: workspaces
-    description: Workspace-level operations
+  - name: github
+    description: GitHub App installation operations
 
 paths:
   # Health Endpoints
@@ -251,9 +251,9 @@ paths:
   /v2/projects.deleteProject:
     $ref: "./spec/paths/v2/projects/deleteProject/index.yaml"
 
-  # Workspace Endpoints
-  /v2/workspaces.installGithub:
-    $ref: "./spec/paths/v2/workspaces/installGithub/index.yaml"
+  # GitHub Endpoints
+  /v2/github.installApp:
+    $ref: "./spec/paths/v2/github/installApp/index.yaml"
 
   # App Endpoints
   /v2/apps.createApp:
