+++ svc/api/openapi/spec/paths/v2/github/installApp/V2GithubInstallAppResponseData.yaml
